tests/wpt/web-platform-tests/IndexedDB/idbtransaction-db-SameObject.html

<!doctype html>
<meta charset=utf-8>
<title>IndexedDB: Verify [SameObject] behavior of IDBTransaction's db attribute</title>
<meta name="help" href="https://w3c.github.io/IndexedDB/#dom-idbtransaction-db">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>
<script>

indexeddb_test(
  (t, db, tx) => {
    const store = db.createObjectStore('store');
    assert_equals(tx.db, tx.db,
                  'Attribute should yield the same object each time');
  },
  (t, db) => {
    const tx = db.transaction('store');
    assert_equals(tx.db, tx.db,
                  'Attribute should yield the same object each time');
    t.done();
  },
  'IDBTransaction.db [SameObject]'
);
</script>
